@@ -8,17 +8,17 @@ namespace NPrng.Generators
         /// Based on "Computationally Easy, Spectrally Good Multipliers for Congruential Pseudorandom Number Generators"
         /// paper by Guy Steele and Sebastiano Vigna
         /// </summary>
-        private const UInt64 Multiplier = 0xfc0072fa0b15f4fd;
+        private const ulong Multiplier = 0xfc0072fa0b15f4fd;
 
         /// <summary>
         /// Based on "The Art of Computer Programming, Volume 2: Seminumerical Algorithms" by Donald E. Knuth
         /// chapter §3.2.1.2, Theorem A.
         /// </summary>
-        private const UInt64 LinearConstant = 34537;  // Prime number not dividing the Multiplier
+        private const ulong LinearConstant = 34537;  // Prime number not dividing the Multiplier
 
-        internal UInt64 CurrentState { get; private set; }
+        internal ulong CurrentState { get; private set; }
 
-        public LinearCongruentGenerator(UInt64 seed)
+        public LinearCongruentGenerator(ulong seed)
         {
             if (seed == 0)
             {
@@ -28,7 +28,7 @@ namespace NPrng.Generators
         }
 
         /// <inheritdoc/>
-        public override Int64 Generate()
+        public override long Generate()
         {
             var currentState = CurrentState;
             unchecked
@@ -36,7 +36,7 @@ namespace NPrng.Generators
                 currentState = Multiplier * currentState + LinearConstant;
             }
             CurrentState = currentState;
-            return (Int64)currentState;
+            return (long)currentState;
         }
     }
 }
