@@ -1,16 +1,15 @@
-using System;
 using System.Runtime.CompilerServices;
 
 namespace NPrng.Generators
 {
     public sealed class Xoshiro256StarStar : AbstractPseudoRandomGenerator
     {
-        internal UInt64 S0 { get; private set; }
-        internal UInt64 S1 { get; private set; }
-        internal UInt64 S2 { get; private set; }
-        internal UInt64 S3 { get; private set; }
+        internal ulong S0 { get; private set; }
+        internal ulong S1 { get; private set; }
+        internal ulong S2 { get; private set; }
+        internal ulong S3 { get; private set; }
 
-        internal Xoshiro256StarStar(UInt64 s0, UInt64 s1, UInt64 s2, UInt64 s3)
+        internal Xoshiro256StarStar(ulong s0, ulong s1, ulong s2, ulong s3)
         {
             S0 = s0;
             S1 = s1;
@@ -20,18 +19,18 @@ namespace NPrng.Generators
 
         public Xoshiro256StarStar(IPseudoRandomGenerator seeder)
         {
-            S0 = (UInt64)seeder.Generate();
-            S1 = (UInt64)seeder.Generate();
-            S2 = (UInt64)seeder.Generate();
-            S3 = (UInt64)seeder.Generate();
+            S0 = (ulong)seeder.Generate();
+            S1 = (ulong)seeder.Generate();
+            S2 = (ulong)seeder.Generate();
+            S3 = (ulong)seeder.Generate();
         }
 
-        public Xoshiro256StarStar(UInt64 seed)
+        public Xoshiro256StarStar(ulong seed)
             : this(new LinearCongruentGenerator(seed))
         { }
 
         /// <inheritdoc/>
-        public override Int64 Generate()
+        public override long Generate()
         {
             unchecked
             {
@@ -43,11 +42,11 @@ namespace NPrng.Generators
                 S0 ^= S3;
                 S2 ^= t;
                 S3 ^= rol64(S3, 45);
-                return (Int64)result;
+                return (long)result;
             }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static UInt64 rol64(UInt64 x, int k) => unchecked((x << k) | (x >> (64 - k)));
+        private static ulong rol64(ulong x, int k) => unchecked((x << k) | (x >> (64 - k)));
     }
 }
