@@ -1,12 +1,12 @@
+using NPrng.Generators;
 using System;
 using System.IO;
-using NPrng.Generators;
 
 namespace NPrng.Serializers
 {
     public sealed class ChaCha20GeneratorSerializer : AbstractPseudoRandomGeneratorSerializer<ChaCha20Generator>
     {
-        private const int BufferSize = 16 * sizeof(UInt32);
+        private const int BufferSize = 16 * sizeof(uint);
 
         /// <inheritdoc/>
         public override int GetExpectedBufferSize(ChaCha20Generator generator) => BufferSize;
@@ -14,12 +14,12 @@ namespace NPrng.Serializers
         /// <inheritdoc/>
         public override ChaCha20Generator ReadFromBuffer(ArraySegment<byte> buffer, out int read)
         {
-            var key = new UInt32[16];
+            var key = new uint[16];
             for (var i = 0; i < 8; i++)
             {
                 buffer = SerializationHelpers.ReadFromBuffer(buffer, out var value);
-                key[2*i+1] = (UInt32)(value & 0xffffffff);
-                key[2*i] = (UInt32)((value >> 32) & 0xffffffff);
+                key[2 * i + 1] = (uint)(value & 0xffffffff);
+                key[2 * i] = (uint)((value >> 32) & 0xffffffff);
             }
             read = BufferSize;
             return new ChaCha20Generator(key);
@@ -28,12 +28,12 @@ namespace NPrng.Serializers
         /// <inheritdoc/>
         public override ChaCha20Generator ReadFromStream(Stream source)
         {
-            var key = new UInt32[16];
+            var key = new uint[16];
             for (var i = 0; i < 8; i++)
             {
                 SerializationHelpers.ReadFromStream(source, out var value);
-                key[2*i+1] = (UInt32)(value & 0xffffffff);
-                key[2*i] = (UInt32)((value >> 32) & 0xffffffff);
+                key[2 * i + 1] = (uint)(value & 0xffffffff);
+                key[2 * i] = (uint)((value >> 32) & 0xffffffff);
             }
             return new ChaCha20Generator(key);
         }
@@ -44,7 +44,7 @@ namespace NPrng.Serializers
             var key = generator.Key;
             for (var i = 0; i < 8; i++)
             {
-                var no = ((UInt64)key[2*i] << 32) + ((UInt64)key[2*i+1]);
+                var no = ((ulong)key[2 * i] << 32) + key[2 * i + 1];
                 buffer = SerializationHelpers.WriteToBuffer(no, buffer);
             }
             return BufferSize;
@@ -56,7 +56,7 @@ namespace NPrng.Serializers
             var key = generator.Key;
             for (var i = 0; i < 8; i++)
             {
-                var no = ((UInt64)key[2*i] << 32) + ((UInt64)key[2*i+1]);
+                var no = ((ulong)key[2 * i] << 32) + key[2 * i + 1];
                 SerializationHelpers.WriteToStream(no, target);
             }
         }
